pyproject.toml
[tool.poetry]
name = "falsisignpy"
version = "0.1.0"
description = ""
authors = ["Stephan Rumswinkel <17520641+srwi@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "4.8.1.78"
Pillow = "9.3.0"
PyMuPDF = "1.21.0"
PySimpleGUI = "4.60.4"

[tool.poetry.group.dev.dependencies]
ruff = "0.1.9"
mypy = "1.7.1"
pytest = "7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".git",
    ".pytest_cache",
    ".ruff_cache",
    ".vscode",
    ".idea",
]
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["ANN", "E", "F", "I"]
ignore = [
    "ANN101",  # self should not be annotated.
    "ANN102"  # cls should not be annotated.
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
